Made fields of ActiveStatus pub(crate) and moved tests to deal with that.
