+++ Cargo.toml
@@ -1,4 +1,4 @@
 [workspace]
-members = ["flash-algo", "flash-algo-test", "parity-robustness-check", "parity-reconstruct"]
+members = ["flash-algo", "parity-robustness-check", "parity-reconstruct"]
 exclude = []
 resolver = "2"
