--- flash-algo-test/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-name = "flash-algo-test"
-version = "0.1.0"
-edition = "2021"
-license = "MIT OR Apache-2.0"
-authors = ["James Munns <james@onevariable.com>"]
-readme = "README.md"
-description = "Flash algorithm for the Cozy Motherboard 2023."
-repository = "https://github.com/Radiator-Labs/cozyboard_2023/tree/master/shared-sw-fw/flash-algo-test"
-keywords = ["no-std", "embedded", "flash-memory"]
-categories = ["Embedded development", "No standard library"]
-
-[badges]
-maintenance = { status = "actively-developed" }
-
-[features]
-# Enable a more robust forward error correction coverage approach.
-#
-# This MAY be incompatible with some other FEC implementations found in the wild.
-force-full-r = ["flash-algo/force-full-r"]
-
-[dependencies]
-crc = "3.2.1"
-flash-algo = { path = "../flash-algo" }
-
-[dev-dependencies]
-insta = "1.39.0"
-tokio = { version = "1.38.0", features = [
-    "macros",
-    "test-util",
-    "rt",
-    "sync",
-    "rt-multi-thread",
-] }
