--- flash-algo-test/rust-toolchain.toml
@@ -1,4 +1,0 @@
-[toolchain]
-channel = "stable"
-# components = ["rustfmt", "rustc-dev"]
-targets = ["thumbv7em-none-eabi"]
