+++ flash-algo/src/tests.rs
@@ -1,0 +1,1 @@
+mod more_real;
