+++ flash-algo/src/testutils.rs
@@ -1,0 +1,5 @@
+#![allow(clippy::std_instead_of_alloc)]
+#![allow(clippy::std_instead_of_core)]
+
+pub mod heap_flash;
+pub mod protocol_test_support;
