+++ flash-algo/test-assets/firmware-001/example.bin
